refactor: move learning modules into learning/ package
Phase 7 of #625 - reorganize project structure:
- learning_orchestrator.py → learning/orchestrator.py
- weather_correlation.py → learning/correlation.py
- Tests moved to tests/learning/
- Updated all imports and mock patches
