+++ pyproject.toml
@@ -42,7 +42,6 @@ select = [
 "custom_components/localshift/computation_engine_lib/price_calculator.py" = ["C901"]
 "custom_components/localshift/computation_engine_lib/slot_builder.py" = ["C901"]
 "custom_components/localshift/computation_engine_lib/solar_utils.py" = ["C901"]
-"custom_components/localshift/state/machine.py" = ["C901"]
 "custom_components/localshift/weather_correlation.py" = ["C901"]
 
 [tool.vulture]
